+++ workspace/apps/server-test/app/utils/trpc.ts
@@ -15,14 +15,18 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import type { FastifyInstance } from "fastify";
-import { register as registerJwtPlugin } from "./jwt";
-import { register as registerCookiePlugin } from "./cookie";
-import { register as registerStaticPlugin } from "./static";
+import {
+    //
+    createTRPCClient,
+    httpBatchLink,
+} from "@trpc/client";
+import type { TTrpcRouter } from "@repo/server/src/routers/trpc/router";
 
-export async function init(fastify: FastifyInstance) {
-    await registerJwtPlugin(fastify); // 注册 JWT 插件
-    await registerCookiePlugin(fastify); // 注册 Cookie 插件
-    await registerStaticPlugin(fastify); // 注册静态文件插件
-    await fastify.after();
-}
+export const trpc = createTRPCClient<TTrpcRouter>({
+    links: [
+        httpBatchLink({
+            url: `/trpc`,
+        }),
+    ],
+});
+export default trpc;