+++ workspace/apps/server/src/contexts/index.ts
@@ -16,3 +16,4 @@
  */
 
 export * from "./test";
+export * from "./session";