+++ workspace/apps/server/src/controllers/login.ts
@@ -36,12 +36,22 @@ import {
     verify,
     type IAuthJwtPayload,
 } from "./../utils/jwt";
-import { Role } from "./../utils/role";
+import {
+    //
+    AccessorRole,
+    AccountRole,
+} from "./../utils/role";
+import { tokens } from "../utils/store";
+
 export interface IAccount {
     id: number; // 账户 ID
     name: string; // 账户名
-    role: Role; // 账户权限
+    role: AccessorRole; // 用户角色
     password: string; // 账户密钥
+    token: {
+        id: number;
+        version: number;
+    }; // 令牌信息
 }
 
 /**
@@ -79,37 +89,98 @@ export const loginMutation = procedure
                 /* 校验用户名是否有效 */
                 const account: IAccount = await (async () => {
                     switch (payload.data.role) {
-                        case Role.Administrator:
-                        case Role.Reviewer:
-                        case Role.Staff: {
+                        case AccountRole.Staff: {
                             const staff = await options.ctx.DB.staff.findUniqueOrThrow({
                                 where: {
                                     name: payload.data.username,
                                     deleted: false,
                                 },
+                                select: {
+                                    id: true,
+                                    name: true,
+                                    role: true,
+                                    password: true,
+                                    token_id: true,
+                                },
                             });
+                            const token = await (async () => {
+                                if (staff.token_id === null) {
+                                    /* 新建一个令牌字段并绑定至员工 */
+                                    const token = await options.ctx.DB.token.create({ data: {} });
+                                    await options.ctx.DB.staff.update({
+                                        where: {
+                                            id: staff.id,
+                                            deleted: false,
+                                        },
+                                        data: {
+                                            token_id: token.id,
+                                        },
+                                    });
+                                    return token;
+                                } else {
+                                    const token = await options.ctx.DB.token.findUniqueOrThrow({
+                                        where: {
+                                            id: staff.token_id,
+                                            deleted: false,
+                                        },
+                                    });
+                                    return token;
+                                }
+                            })();
                             return {
                                 id: staff.id,
                                 name: staff.name,
                                 role: staff.role,
                                 password: staff.password,
+                                token,
                             };
                         }
 
-                        case Role.User:
-                        case Role.Visitor:
+                        case AccountRole.User:
                         default: {
                             const user = await options.ctx.DB.user.findUniqueOrThrow({
                                 where: {
                                     name: payload.data.username,
                                     deleted: false,
                                 },
+                                select: {
+                                    id: true,
+                                    name: true,
+                                    password: true,
+                                    token_id: true,
+                                },
                             });
+                            const token = await (async () => {
+                                if (user.token_id === null) {
+                                    /* 新建一个令牌字段并绑定至用户 */
+                                    const token = await options.ctx.DB.token.create({ data: {} });
+                                    await options.ctx.DB.user.update({
+                                        where: {
+                                            id: user.id,
+                                            deleted: false,
+                                        },
+                                        data: {
+                                            token_id: token.id,
+                                        },
+                                    });
+                                    return token;
+                                } else {
+                                    /* 获取对应的令牌 */
+                                    const token = await options.ctx.DB.token.findUniqueOrThrow({
+                                        where: {
+                                            id: user.token_id,
+                                            deleted: false,
+                                        },
+                                    });
+                                    return token;
+                                }
+                            })();
                             return {
                                 id: user.id,
                                 name: user.name,
-                                role: Role.User,
+                                role: AccessorRole.User,
                                 password: user.password,
+                                token,
                             };
                         }
                     }
@@ -126,6 +197,22 @@ export const loginMutation = procedure
                     Buffer.from(account.password, "hex"),
                 );
                 if (response.toString("hex") === options.input.response) {
+                    // 用户名与密码正确
+
+                    /* 更新令牌版本 */
+                    account.token.version++;
+                    tokens.set(account.token.id, account.token.version);
+                    await options.ctx.DB.token.update({
+                        where: {
+                            id: account.token.id,
+                            deleted: false,
+                        },
+                        data: {
+                            version: account.token.version,
+                        },
+                    });
+
+                    /* 签发新令牌 */
                     const payload: IAuthJwtPayload = {
                         data: {
                             account: {
@@ -133,7 +220,12 @@ export const loginMutation = procedure
                                 role: account.role,
                                 username: account.name,
                             },
+                            token: {
+                                id: account.token.id,
+                                version: account.token.version,
+                            },
                         },
+                        jti: account.token.id,
                     };
                     // REF: https://www.npmjs.com/package/@fastify/jwt?activeTab=readme#example-using-cookie
                     const token = options.ctx.S.jwt.sign(payload);
@@ -144,12 +236,12 @@ export const loginMutation = procedure
                         {
                             // REF: https://developer.mozilla.org/zh-CN/docs/Web/HTTP/Cookies
                             // domain: "your.domain", // 默认为同站 (不包括子域名), 设置后将包含子域名
-                            // path: "/", // 将在指定路径下发送该 Cookie
                             // secure: true, // 通过被 HTTPS 协议加密过的请求发送
-                            sameSite: "lax", // 用户从其他站点导航到 Cookie 的源站点时也发送 Cookie
+                            path: "/", // 将在指定路径下发送该 Cookie
+                            sameSite: "lax", // 用户从其他站点导航到 Cookie 的源站点时发送 Cookie
                             httpOnly: true, // 阻止通过 document.cookie 访问该 Cookie
                             expires: options.input.stay // Cookie 有效期
-                                ? new Date(options.ctx.S.jwt.decode<any>(token).exp) // Cookie 有效期与 JWT 一致
+                                ? new Date(options.ctx.S.jwt.decode<any>(token).exp * 1_000) // Cookie 有效期与 JWT 一致
                                 : undefined, // 会话 Cookie (关闭会话时失效)
                         },
                     );