+++ workspace/apps/server/tests/trpc/login.test.ts
@@ -31,58 +31,42 @@ import {
 } from "@repo/utils/crypto";
 
 import trpc from ".";
-import {
-    //
-    Role,
-    str2role,
-} from "./../../src/utils/role";
+import { AccessorRole } from "./../../src/utils/role";
 
 type TRole = Parameters<typeof trpc.auth.challenge.query>[0]["role"];
 
-interface IUserAccount {
+interface IAccount {
     username: string;
     passphrase: string;
     role: TRole;
-    _role: Role;
+    _role: AccessorRole;
 }
 
 describe("/trpc/account/login", () => {
-    const accounts: IUserAccount[] = [
+    const accounts: IAccount[] = [
         {
             username: "admin",
             passphrase: "admin",
-            role: "administrator",
-            _role: Role.Administrator,
-        },
-        {
-            username: "reviewer",
-            passphrase: "reviewer",
-            role: "reviewer",
-            _role: Role.Reviewer,
+            role: "staff",
+            _role: AccessorRole.Administrator,
         },
         {
             username: "reviewer",
             passphrase: "reviewer",
             role: "staff",
-            _role: Role.Reviewer,
+            _role: AccessorRole.Reviewer,
         },
         {
             username: cuid.createId(),
             passphrase: cuid.createId(),
             role: "user",
-            _role: Role.User,
-        },
-        {
-            username: cuid.createId(),
-            passphrase: cuid.createId(),
-            role: "visitor",
-            _role: Role.User,
+            _role: AccessorRole.User,
         },
         {
             username: cuid.createId(),
             passphrase: cuid.createId(),
             role: undefined,
-            _role: Role.User,
+            _role: AccessorRole.User,
         },
     ];
     accounts.forEach((account) => {